# gen/openapi/common/models.gen.go
// Package common provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package common

import (
	"time"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error string `json:"error"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    string `json:"email"`
	Password string `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	Token *string `json:"token,omitempty"`
	User  *User   `json:"user,omitempty"`
}

// SignupRequest defines model for SignupRequest.
type SignupRequest struct {
	Email    string `json:"email"`
	Name     string `json:"name"`
	Password string `json:"password"`
}

// SuccessResponse defines model for SuccessResponse.
type SuccessResponse struct {
	Message string `json:"message"`
}

// Todo defines model for Todo.
type Todo struct {
	Content   *string    `json:"content,omitempty"`
	CreatedAt *time.Time `json:"created_at,omitempty"`
	Id        *uint      `json:"id,omitempty"`
	Title     *string    `json:"title,omitempty"`
	UpdatedAt *time.Time `json:"updated_at,omitempty"`
	UserId    *int       `json:"user_id,omitempty"`
}

// TodoCreateRequest defines model for TodoCreateRequest.
type TodoCreateRequest struct {
	Content string `json:"content"`
	Title   string `json:"title"`
	UserId  uint   `json:"user_id"`
}

// TodoList defines model for TodoList.
type TodoList = []Todo

// TodoUpdateRequest defines model for TodoUpdateRequest.
type TodoUpdateRequest struct {
	Content string `json:"content"`
	Title   string `json:"title"`
}

// User defines model for User.
type User struct {
	Email *string `json:"email,omitempty"`
	Id    *int    `json:"id,omitempty"`
	Name  *string `json:"name,omitempty"`
}

// GetTodoListParams defines parameters for GetTodoList.
type GetTodoListParams struct {
	UserId *uint `form:"userId,omitempty" json:"userId,omitempty"`
}

// PostAuthLoginJSONRequestBody defines body for PostAuthLogin for application/json ContentType.
type PostAuthLoginJSONRequestBody = LoginRequest

// PostAuthSignupJSONRequestBody defines body for PostAuthSignup for application/json ContentType.
type PostAuthSignupJSONRequestBody = SignupRequest

// CreateTodoJSONRequestBody defines body for CreateTodo for application/json ContentType.
type CreateTodoJSONRequestBody = TodoCreateRequest

// UpdateTodoJSONRequestBody defines body for UpdateTodo for application/json ContentType.
type UpdateTodoJSONRequestBody = TodoUpdateRequest