# infrastructure/middleware/jwt.go
package middleware

import (
	"fmt"
	"net/http"
	"strings"
	"todo-golang/config"

	"github.com/golang-jwt/jwt/v5"
	"github.com/labstack/echo/v4"
)

func JWTMiddleware(next echo.HandlerFunc) echo.HandlerFunc {
	return func(c echo.Context) error {
		if c.Path() == "/api/v1/auth/login" || c.Path() == "/api/v1/auth/signup" {
			return next(c)
		}

		authHeader := c.Request().Header.Get("Authorization")
		fmt.Println("Authorization Header:", authHeader)
		if authHeader == "" {
			return echo.NewHTTPError(http.StatusUnauthorized, "Missing Authorization Header")
		}

		tokenStr := strings.TrimPrefix(authHeader, "Bearer ")

		token, err := jwt.Parse(tokenStr, func(token *jwt.Token) (interface{}, error) {
			if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
				return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
			}

			return []byte(config.AppConf.Auth.JWT_SECRET), nil
		})

		if err != nil || !token.Valid {
			return echo.NewHTTPError(http.StatusUnauthorized, "Invalid Token")
		}

		claims, ok := token.Claims.(jwt.MapClaims)
		if !ok {
			return echo.NewHTTPError(http.StatusUnauthorized, "Invalid Token Claims")
		}

		c.Set("user", claims["email"])

		return next(c)
	}
}